strings.xml: make it easier to translate
"Conflict Notes" may be translated as "Notes about the conflict", so I changed it to "Conflicting Notes".

"About" is too broad (in Russian it's just "О"), so I added "About the app".

"Check the GitHub" translated as "check for errors", so I replaced it to just "See".

"Credits" may be translated as money credits, so I changed to "Authors".

"Submit an issue" is simplified because the "issue" always translated as a "problem" and the "bug" is translated similarly